requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Jinja2==3.1.2
WTForms==3.0.1
mysql-connector-python==2.2.9
